refactor: #681 src/lib/api.ts 모놀리스를 도메인별 모듈로 분리
- src/lib/api.ts(1596 lines)를 src/lib/api/ 디렉터리 아래 도메인별 모듈로 분리
- core.ts: ApiClient 클래스, 401 인터셉터(/auth/ skip 규칙 유지), request helpers, 공통 타입(PaginatedResponse, ListResponse, UploadedFile, RequestOptions)
- 도메인 모듈: products, cart, orders, payments, auth, users, shipping, pages, navigation, reviews, wishlist, coupons, notices, faqs, inquiries, promotions, banners, settings, collections, archives, journals
- admin 하위 모듈: admin/{categories, products, orders, members, dashboard, pages, navigation, faqs, inquiries, settings, collections, archives, journals}
- src/lib/api/index.ts 배럴로 re-export — 모든 @/lib/api 임포트 경로 호환 유지
- 기존 테스트 통과(407 tests), lint/build 성공

Closes #681
